Update: For docker imaging, chnaged directories and codes.
